Added authorization

+++ AptekaWebAPI/Controllers/Cart.cs
@@ -25,7 +25,7 @@ namespace AptekaWebAPI.Controllers
         public ActionResult AddById([FromHeader] string token, [FromBody] AddToCartDTO dto)
         {
             var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
-            if (activeUser == null) throw new Exception();
+            if (activeUser == null || activeUser.Id < 4) throw new Exception();
             _service.AddById(activeUser.Id, dto);
             return Ok();
         }
@@ -33,6 +33,9 @@ namespace AptekaWebAPI.Controllers
         [HttpDelete("remove/{id}")]
         public ActionResult RemoveById([FromHeader] string token, [FromRoute] int id)
         {
+            var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
+            if (activeUser == null || activeUser.Id < 4) throw new Exception();
+
             _service.RemoveById(id);
             return Ok();
         }
@@ -40,6 +43,9 @@ namespace AptekaWebAPI.Controllers
         [HttpPut("update")]
         public ActionResult Modify([FromHeader] string token, [FromBody] AddToCartDTO dto)
         {
+            var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
+            if (activeUser == null || activeUser.Id < 4) throw new Exception();
+
             _service.Modify(dto);
             return Ok();
         }
@@ -48,7 +54,7 @@ namespace AptekaWebAPI.Controllers
         public ActionResult GetAllFavourites([FromHeader] string token)
         {
             var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
-            if (activeUser == null) throw new Exception();
+            if (activeUser == null || activeUser.Id < 4) throw new Exception();
 
             var allProducts = _service.GetAll(activeUser.Id);
             return Ok(allProducts);
@@ -57,6 +63,8 @@ namespace AptekaWebAPI.Controllers
         [HttpGet("{id}")]
         public ActionResult GetFavouriteByID([FromHeader] string token, [FromRoute] int id)
         {
+            var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
+            if (activeUser == null || activeUser.Id < 4) throw new Exception();
             var selectedProduct = _service.GetByID(id);
             return Ok(selectedProduct);
         }
