+++ AptekaWebAPI/Controllers/PharmacyAdmin.cs
@@ -1,5 +1,6 @@
 ﻿using AptekaWebAPI.Properties.DTOs;
 using AptekaWebAPI.Services;
+using AptekaWebAPI.Tokens;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
@@ -23,6 +24,8 @@ namespace AptekaWebAPI.Controllers
         [HttpPost("addProduct")]
         public ActionResult AddNewProduct([FromHeader] string token, [FromBody] CreateProductDTO dto)
         {
+            var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
+            if (activeUser == null || activeUser.Id > 3) throw new Exception();
             _service.AddNewProduct(dto);
             return Ok();
         }
@@ -30,6 +33,8 @@ namespace AptekaWebAPI.Controllers
         [HttpGet("allProducts")]
         public ActionResult<IEnumerable<ProductDTO>> GetAllProducts([FromHeader] string token)
         {
+            var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
+            if (activeUser == null || activeUser.Id > 3) throw new Exception();
             var products = _service.GetAll();
             return Ok(products);
         }
@@ -37,6 +42,9 @@ namespace AptekaWebAPI.Controllers
         [HttpGet("{id}")]
         public ActionResult<ProductDTO> GetProductById([FromHeader] string token, [FromRoute] int id)
         {
+            var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
+            if (activeUser == null || activeUser.Id > 3) throw new Exception();
+
             var product = _service.GetById(id);
             return Ok(product);
         }
@@ -45,6 +53,9 @@ namespace AptekaWebAPI.Controllers
         [HttpPut("update")]
         public ActionResult Modify([FromHeader] string token,[FromBody] UpdateProductDTO dto)
         {
+            var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
+            if (activeUser == null || activeUser.Id > 3) throw new Exception();
+
             _service.Modify(dto);
             return Ok();
         }
@@ -52,6 +63,8 @@ namespace AptekaWebAPI.Controllers
         [HttpDelete("remove/{id}")]
         public ActionResult Delete([FromHeader] string token, [FromRoute] int id)
         {
+            var activeUser = new ActiveUsers().GetAllLoginedUsers().Where(x => x.Token == token).FirstOrDefault();
+            if (activeUser == null || activeUser.Id > 3) throw new Exception();
             _service.Delete(id);
             return Ok();
         }
